Adding missing data files and updating state_pension_amount_calculator
The 'publish' rake task doesn't know about these additional data files so this adds them. Updaing the calculator file to point to v1 files.
