+++ lib/data/state_pension_query.rb
@@ -15,21 +15,29 @@ end
 class StatePensionQuery < Struct.new(:dob, :gender)
   def self.find(dob, gender)
     state_pension_query = new(dob, gender)
-    result = state_pension_query.run
-    state_pension_query.account_for_leap_year(result.pension_date)
+    state_pension_query.find_date
   end
 
-  def run
-    state_pension_dates.find {|p| p.match?(dob, gender)}
+  def find_date
+    static_result = run(pension_dates_static)
+    if static_result
+      static_result.pension_date
+    else
+      result = run(pension_dates_dynamic)
+      adjust_when_dob_is_29february(result.pension_date)
+    end
+  end
+
+  def run(pension_dates)
+    pension_dates.find {|p| p.match?(dob, gender)}
   end
 
   # Handle the case where the person's d.o.b. is 29th Feb
-  # on a leap year and that the pension elligibility date falls
-  # on a non-leap year. ActiveSupport will helpfully adjust the
-  # date back to 28th Feb, but the calculation should forward to
-  # the 1st March according to DWP rules.
-  #
-  def account_for_leap_year(date)
+  # on a leap year and that the pension eligibility date falls
+  # on a non-leap year and when dob falls into the dynamic date group.
+  # ActiveSupport will adjust the date, but the calculation
+  # should adjust to the 1st March according to DWP rules.
+  def adjust_when_dob_is_29february(date)
     if leap_year_date?(dob) and !leap_year_date?(date)
       date += 1
     end
@@ -38,10 +46,6 @@ class StatePensionQuery < Struct.new(:dob, :gender)
 
   def leap_year_date?(date)
     date.month == 2 and date.day == 29
-  end
-
-  def state_pension_dates
-    pension_dates_static + pension_dates_dynamic
   end
 
   def pension_dates_dynamic
@@ -66,6 +70,6 @@ class StatePensionQuery < Struct.new(:dob, :gender)
   end
 
   def pension_dates_static
-    @@pension_dates_static ||= YAML.load_file(Rails.root.join("lib", "data", "state_pension_dates.yml"))
+    pension_dates_static ||= YAML.load_file(Rails.root.join("lib", "data", "state_pension_dates.yml"))
   end
 end
